rag_support_bot/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
langgraph>=0.1.0
chromadb>=0.4.0
pypdf>=3.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
reportlab>=4.0.0
groq>=0.4.0
